Add Multipass Support
Add support for Multipass. Tugboat and FishyWebRTC are now both supported.
Add Client.jslib which contains ClientWebGLQuit function for closing the WebGL clients.
Add UNITY_WEBGL define checks for GA4, and Client transport initialization.
Add Server.RegisterBroadcast and Server.UnregisterBroadcast. These can be used to add additional functionality to the broadcast registrar.
Add Destroying function to ServerBehaviour. This function is called when an object is set to be destroyed.
Add Server port configuration support when creating new cfg files. Default ports are 7770 for Login,  7780 for World, 7781 for Scene servers.
Add Multipass support to ServerWindowTitleUpdater.
Fix Server systems are now initialized fully in InitializeOnce, this includes broadcast registrations.
Fix null reference exception in UIChat and ChatSystem related to Character.LastChatMessage.
Fix Multipass GetPort function. It will now return the first transport port when called instead of just 0.
Remove redundant exception logging in FishyWebRTC.
Launch Local Game Server.bat is now copied to WebGL client builds. This can be used to initialize a python web server quickly for testing purposes or if you want to host the client locally.
AreaBuff, AreaDamage, and AreaHeal hit events now have a sweep limit of 512. This will eventually be used for deterministic entity targetting.

+++ FishMMO-Unity/Assets/Scripts/Server/World/SceneServer/SceneServerSystem.cs
@@ -31,31 +31,20 @@ namespace FishMMO.Server
 
 		public override void InitializeOnce()
 		{
+			using var dbContext = Server.NpgsqlDbContextFactory.CreateDbContext();
+			if (dbContext == null)
+			{
+				throw new UnityException("Failed to get dbContext.");
+			}
+
 			if (ServerManager != null &&
+				Server != null &&
 				Server.NetworkManager.SceneManager != null)
 			{
 				Server.NetworkManager.SceneManager.OnLoadEnd += SceneManager_OnLoadEnd;
 				//Server.NetworkManager.SceneManager.OnUnloadEnd += SceneManager_OnUnloadEnd;
 				ServerManager.OnServerConnectionState += ServerManager_OnServerConnectionState;
-			}
-			else
-			{
-				enabled = false;
-			}
-		}
 
-		private void ServerManager_OnServerConnectionState(ServerConnectionStateArgs args)
-		{
-			serverState = args.ConnectionState;
-
-			using var dbContext = Server.NpgsqlDbContextFactory.CreateDbContext();
-			if (dbContext == null)
-			{
-				return;
-			}
-
-			if (args.ConnectionState == LocalConnectionState.Started)
-			{
 				if (Server.TryGetServerIPAddress(out ServerAddress server) &&
 					ServerBehaviour.TryGet(out CharacterSystem characterSystem))
 				{
@@ -68,15 +57,35 @@ namespace FishMMO.Server
 					}
 				}
 			}
-			else if (args.ConnectionState == LocalConnectionState.Stopped)
+			else
 			{
-				if (Server.Configuration.TryGetString("ServerName", out string name))
+				enabled = false;
+			}
+		}
+
+		public override void Destroying()
+		{
+			using var dbContext = Server.NpgsqlDbContextFactory.CreateDbContext();
+			if (dbContext == null)
+			{
+				throw new UnityException("Failed to get dbContext.");
+			}
+
+			if (ServerManager != null)
+			{
+				if (Server != null &&
+					Server.Configuration.TryGetString("ServerName", out string name))
 				{
 					Debug.Log("Scene Server System: Removing Scene Server: " + id);
 					SceneServerService.Delete(dbContext, id);
 					LoadedSceneService.Delete(dbContext, id);
 				}
 			}
+		}
+
+		private void ServerManager_OnServerConnectionState(ServerConnectionStateArgs args)
+		{
+			serverState = args.ConnectionState;
 		}
 
 		void LateUpdate()
@@ -121,18 +130,6 @@ namespace FishMMO.Server
 					}
 				}
 				nextPulse -= Time.deltaTime;
-			}
-		}
-
-		private void OnApplicationQuit()
-		{
-			if (Server != null && Server.NpgsqlDbContextFactory != null &&
-				serverState != LocalConnectionState.Stopped)
-			{
-				using var dbContext = Server.NpgsqlDbContextFactory.CreateDbContext();
-				Debug.Log("Scene Server System: Removing Scene Server: " + id);
-				SceneServerService.Delete(dbContext, id);
-				LoadedSceneService.Delete(dbContext, id);
 			}
 		}
